spire: provide basic virt net autoconfig commands
In order to facilitate launching local clusters, add a set of commands
to spire that can bring the local virtual network up and down. This
includes bringing up and down bridges and tap devices, adding and
removing iptables rules and /etc/hosts entries, and automatically
computing the correct actions to take from the configuration.